Minor changes and fixes

+++ resources/views/user_views/messages/tables/read_messages_table.blade.php
@@ -31,12 +31,14 @@
                     <i class="fa-solid fa-eye fs-5 me-2"></i>
                 </a>
             @endif
-            {!! Form::open(['route' => ['messages.destroy', [$prefix, $readMessage->message->id]], 'method' => 'delete']) !!}
-                <button type="submit" class="mark-as-read-button" title="{{ __('buttons.deleteMessage') }}"
-                        onclick="return confirm('{{ __('names.areYouSureDeleteMessage') }}')">
-                    <i class="fa-solid fa-trash-can fs-5"></i>
-                </button>
-            {!! Form::close() !!}
+            @if ($readMessage->message->sender_id == auth()->user()->id)
+                {!! Form::open(['route' => ['messages.destroy', [$prefix, $readMessage->message->id]], 'method' => 'delete']) !!}
+                    <button type="submit" class="mark-as-read-button" title="{{ __('buttons.deleteMessage') }}"
+                            onclick="return confirm('{{ __('names.areYouSureDeleteMessage') }}')">
+                        <i class="fa-solid fa-trash-can fs-5"></i>
+                    </button>
+                {!! Form::close() !!}
+            @endif
         </div>
     </div>
 @empty
